pyproject.toml
[tool.poetry]
name = "pdf-parser"
version = "0.1.0"
description = "Full-featured PDF document analysis with AI extraction, database persistence, and batch processing."
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "pdf_parser", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"

# Core
streamlit = "^1.28"
python-dotenv = "^1.0"
pydantic = "^2.5"
sqlalchemy = "^2.0"
alembic = "^1.13"

# PDF processing libraries (fallback chain)
PyPDF2 = "^3.0"
pdfplumber = "^0.10"
pymupdf = "^1.23"
pdf2image = "^1.17"

# Image processing
Pillow = "^10.0"

# OpenAI & HTTP
openai = "^1.6"
httpx = "^0.25"
aiofiles = "^23.0"

# Data & validation
pandas = "^2.1"
numpy = "^1.25"

# Logging & utils
loguru = "^0.7"
tqdm = "^4.66"
click = "^8.1"
python-multipart = "^0.0.6"

# Security
cryptography = "^41.0"
langfuse = "^3.2.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-asyncio = "^0.21"
pytest-cov = "^4.1"
pytest-mock = "^3.12"
black = "^23.9"
flake8 = "^6.1"
mypy = "^1.7"
pre-commit = "^3.5"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
